Split signin and signup into different pages
