+++ client/src/views/SignUp.vue
@@ -1,0 +1,137 @@
+<template>
+  <v-layout justify-center>
+    <v-flex xs4>
+      <v-card>
+        <v-card-title class="justify-center" primary-title>
+          <h2>Sign Up</h2>
+        </v-card-title>
+        <v-card-text>
+        <p style="color: red">{{ signUpError }}</p>
+        <v-stepper class="elevation-0" v-model="signUpStep">
+          <v-stepper-content step="1">
+            <v-btn color="red" @click="signUpWithGoogle" dark>Googleでアカウント作成</v-btn>
+            <br />
+            <v-btn color="light-blue" @click="signUpWithTwitter" dark>Twitterでアカウント作成</v-btn>
+          </v-stepper-content>
+
+          <v-stepper-content step="2">
+            <v-container>
+              <form>
+                <v-flex>
+                  <v-text-field
+                    v-model="form.name"
+                    label="ユーザーID"
+                    append-outer-icon="check"
+                  />
+                  <v-text-field
+                    v-model="form.display_name"
+                    label="表示される名前"
+                  />
+                </v-flex>
+                <v-avatar color="orange" size="32px">
+                  <v-img :src="form.picture" />
+                </v-avatar>
+                <v-btn depressed>アイコンをアップロード</v-btn>
+
+                <br />
+
+                <v-btn color="success" @click="signUp">送信</v-btn>
+                <v-btn depressed @click="signUpStep --;">キャンセル</v-btn>
+              </form>
+            </v-container>
+          </v-stepper-content>
+        </v-stepper>
+        
+        <br />
+        <router-link to="/signin">アカウントを持っている場合はこちら</router-link>
+        </v-card-text>
+      </v-card>
+    </v-flex>
+  </v-layout>
+</template>
+
+<script>
+import axios from 'axios';
+import sdk from '@/app/sdk';
+
+export default {
+  props: [ 'signup' ],
+  data () {
+    return {
+      form: {
+        name: '',
+        display_name: '',
+        picture: '',
+      },
+      logins: {},
+      signUpStep: 1,
+      signUpError: '',
+    };
+  },
+  methods: {
+    async toDashboard ({ id_token, user }) {
+      localStorage.setItem('id_token', id_token);
+      localStorage.setItem('user', JSON.stringify(user));
+      this.$router.push('/dashboard');
+    },
+    async signUpWithGoogle () {
+      const user = await this.$gAuth.signIn();
+      const profile = user.getBasicProfile();
+
+      this.form = {
+        name: profile.getId(),
+        display_name: profile.getName(),
+        picture: profile.getImageUrl(),
+      };
+      this.logins = {
+        google: user.getAuthResponse().id_token
+      };
+
+      this.signUpStep ++;
+    },
+    async signUpWithTwitter () {
+      const twitterAuthURL = await axios.post(`${process.env.VUE_APP_API_ENDPOINT}/auth/twitter`);
+      location.href = twitterAuthURL.data;
+    },
+    async signUpWithTwitterAfter (token) {
+      const result = (await axios.get(`${process.env.VUE_APP_API_ENDPOINT}/auth/twitter?oauth_token=${token.oauth_token}&oauth_verifier=${token.oauth_verifier}`)).data;
+      const account = result.account;
+      console.log(account)
+      console.log(token)
+
+      this.form = {
+        name: account.screen_name,
+        display_name: account.screen_name,
+        picture: account.profile_image_url,
+      }
+      this.logins = {
+        twitter: result.credential,
+      };
+
+      this.signUpStep ++;
+    },
+    async signUp () {
+      try {
+        const result = (await sdk.signUp({
+          form: this.form,
+          logins: this.logins,
+        })).data;
+        await this.toDashboard(result);
+      } catch (err) {
+        this.signUpError = 'SignUpError';
+        this.$router.push('/signup');
+        this.signUpStep = 1;
+        return;
+      }
+    }
+  },
+  mounted () {
+    // twitter-callback
+    // Continue to signUpWithTwitter
+    if (this.$route.path.startsWith('/signup/twitter-callback')) {
+      this.signUpWithTwitterAfter(this.$route.query);
+    }
+},
+}
+</script>
+
